There seem to be more like DIRSERVER-1329
"The parameter annotation identifies the variable as a mojo parameter. The default-value parameter of the annotation defines the default value for the variable. This value can include expressions which reference the project, such as "${project.version}" (more can be found in the "Parameter Expressions" document). The expression parameter can be used to allow configuration of the mojo parameter from the command line by referencing a system property that the user sets via the -D option."

git-svn-id: https://svn.apache.org/repos/asf/directory/daemon/trunk@753616 13f79535-47bb-0310-9956-ffa450edef68